modified implementation of stack to store address

+++ Chapter 10/Stack/stack.h
@@ -5,14 +5,14 @@
 
 typedef struct stack
 {
-	long* stack;
+	unsigned long long* stack;
 	int top;
 	int size;
 } stack;
 
 bool is_stack_empty(stack* s);
-void push(stack* s, long x);
-long pop(stack* s);
+void push(stack* s, unsigned long long x);
+unsigned long long pop(stack* s);
 bool create_stack(stack* s, int size);
 void clear_stack(stack* s);
